LPS-127341 Create frontend-js-a11y-web module
