+++ modules/apps/frontend-js/frontend-js-a11y-web/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Frontend JS A11y Web
+Bundle-SymbolicName: com.liferay.frontend.js.a11y.web
+Bundle-Version: 1.0.0
+Web-ContextPath: /frontend-js-a11y-web
