+++ modules/apps/frontend-js/frontend-js-a11y-web/package.json
@@ -1,0 +1,19 @@
+{
+	"dependencies": {
+		"@clayui/icon": "3.1.0",
+		"@clayui/popover": "3.28.0",
+		"@clayui/list": "3.28.0",
+		"axe-core": "4.2.0",
+		"react": "16.12.0",
+		"react-dom": "16.12.0"
+	},
+	"main": "index.js",
+	"name": "@liferay/frontend-js-a11y-web",
+	"scripts": {
+		"build": "liferay-npm-scripts build",
+		"checkFormat": "liferay-npm-scripts check",
+		"format": "liferay-npm-scripts fix",
+		"test": "liferay-npm-scripts test"
+	},
+	"version": "1.0.0"
+}
