+++ modules/apps/frontend-js/frontend-js-a11y-web/src/main/resources/META-INF/resources/A11yChecker.ts
