+++ modules/apps/frontend-js/frontend-js-a11y-web/src/main/resources/META-INF/resources/index.ts
@@ -12,22 +12,4 @@
  * details.
  */
 
-import type {CheckResult, ImpactValue} from 'axe-core';
-import type {A11yCheckerOptions} from '../components/A11y/A11yChecker';
-export declare type Violation = {
-	all: Array<CheckResult>;
-	any: Array<CheckResult>;
-	help: string;
-	helpUrl: string;
-	id: string;
-	impact?: ImpactValue;
-};
-declare type Violations = {
-	modifyIndex: number;
-	target: string;
-	violations: Array<Violation>;
-};
-export default function useA11y(
-	props: Omit<A11yCheckerOptions, 'callback'>
-): Violations[];
-export {};
+export {A11y} from './A11y';
