+++ modules/apps/frontend-js/frontend-js-a11y-web/types/src/main/resources/META-INF/resources/A11y.d.ts
@@ -16,6 +16,6 @@
 
 import './A11y.scss';
 import type {A11yCheckerOptions} from './A11yChecker';
-export default function A11y(
+export declare function A11y(
 	props: Omit<A11yCheckerOptions, 'callback'>
 ): JSX.Element[] | null;
