+++ modules/apps/frontend-js/frontend-js-a11y-web/types/src/main/resources/META-INF/resources/A11yChecker.d.ts
