+++ modules/apps/frontend-js/frontend-js-a11y-web/types/src/main/resources/META-INF/resources/index.d.ts
@@ -12,4 +12,4 @@
  * details.
  */
 
-export {default as A11y} from './A11y';
+export {A11y} from './A11y';
